+++ packages/hummingbird/lib/theme/design.js
@@ -32,7 +32,7 @@ export default ({
   // Footer
   footerBackgroundColor,
   footerTextColor,
-  // Articles/pages
+  // Articles/Pages
   chapterBackgroundColor,
   chapterItemColor,
   chapterItemNumberColor,
@@ -44,6 +44,7 @@ export default ({
   blockquoteBackgroundColor,
   pullquoteColor,
   pullquoteBackgroundColor,
+  testimonialBackgroundColor,
   // Landing Section
   landingSectionTitleColor = '#FFFFFF',
   landingSectionTitleBackgroundColor = 'transparent',
@@ -53,9 +54,13 @@ export default ({
   // Forms
   formBackgroundColor,
   formTextColor,
+  formLinkTextColor,
+  formLinkLineColor,
   // Campaigns
   campaignFormBackgroundColor,
   campaignFormTextColor,
+  campaignFormLinkTextColor,
+  campaignFormLinkLineColor,
   counterBackgroundColor,
   counterTextColor,
 
@@ -162,6 +167,7 @@ export default ({
   blockquoteBackgroundColor: blockquoteBackgroundColor || backgroundColor,
   pullquoteColor: pullquoteColor || primaryColor,
   pullquoteBackgroundColor: pullquoteBackgroundColor || backgroundColor,
+  testimonialBackgroundColor: testimonialBackgroundColor || elementBackgroundColor,
   // Landing Section
   landingSectionTitleColor,
   landingSectionTitleBackgroundColor,
@@ -177,6 +183,15 @@ export default ({
       colors: { light: textColor, dark: textColorDark },
       threshold: contrastLuminanceThreshold,
     }),
+  formLinkTextColor:
+    formLinkTextColor ||
+    formTextColor ||
+    getContrastColor({
+      backgroundColor: formBackgroundColor || primaryColor,
+      colors: { light: textColor, dark: textColorDark },
+      threshold: contrastLuminanceThreshold,
+    }),
+  formLinkLineColor: secondaryColor || formLinkLineColor,
   // Campaigns
   campaignFormBackgroundColor: campaignFormBackgroundColor || formBackgroundColor || primaryColor,
   campaignFormTextColor:
@@ -187,6 +202,17 @@ export default ({
       colors: { light: textColor, dark: textColorDark },
       threshold: contrastLuminanceThreshold,
     }),
+  campaignFormLinkTextColor:
+    campaignFormLinkTextColor ||
+    formLinkTextColor ||
+    campaignFormTextColor ||
+    formTextColor ||
+    getContrastColor({
+      backgroundColor: formBackgroundColor || primaryColor,
+      colors: { light: textColor, dark: textColorDark },
+      threshold: contrastLuminanceThreshold,
+    }),
+  campaignFormLinkLineColor: campaignFormLinkLineColor || formLinkLineColor || secondaryColor,
   counterBackgroundColor: counterBackgroundColor || backgroundColor,
   counterTextColor:
     counterTextColor ||