+++ packages/hummingbird/src/components/Navigation/Navigation.js
@@ -179,7 +179,7 @@ class Navigation extends Component {
             <Logo />
           </Link>
           {children}
-          {hideMenu ? null : (
+          {hideMenu || !items || !items.length ? null : (
             <MenuButton active={visible} onClick={this.toggleNav}>
               <Burger
                 active={visible}